Start to add xml <summary docs to reduce the number of warnings
